Used code for ESP32-WROVER Camera WiFi module
March 07, 2023
PROGRESS WEEK 2-7 MARCH
I started by gathering the equippment for my project:

Mr. Purav Shah provided me with RaspberryPi 2, Ras Pi HQ camera as well as an Arduino Kit - just had another meet today (7th March) and ordered lenses for the RasPi HQ camera so I might incorporate that in the final project as well as the current wifi cam.
Mrs. Ramona Trestian lent me a Samsung Gear360 camera and a Raspberry 3 which compared to the 2, it has better CPU and also has built-in WiFi and Bluetooth which is perfect for my project.
I also purchased

2x drive controller board (L298N)
4x wheels
4x motors
ESP32 Camera with built-in WiFi
Throughout this week I have made some improvements on the hardware side, better defining the aim and the objectives as well as the problems that are on the way.

I managed to

get the motor work as I ordered the powering interface of the motors (L298N).
make the motors work via the IR remote
get the wifi cam connected
get wifi cam live feed on local host
found a protocol (RSTP) that can help me get the live stream from local to remote -- https://github.com/aler9/rtsp-simple-server/
get wifi cam live feed on remove host https://mdxplorer.gymvibeiasi.ro/
the live feed needed a lot of research but mainly it has to do with port forwarding on the home router as well as having the live feed encrypted and working via a TLS (secure) transmission (HTTPS)
make the whole livesreaming process start on boot as well as getting and sending the footage over internet only when there is a person accessing the livestream (on demand service) -- this saves processing power, energy and internet bandwidth and since it's a service it will always restart if crashes

https://andreipaultocu.blogspot.com/2023/03/progress-week-2-7-march.html